+++ inc/functions/security-assessment.php
@@ -11,11 +11,11 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         $Status = $UserInfo['Status'];
         $Error = $UserInfo['Error'];
     } else {
-        // echo json_encode(array(
-        //     'Status' => 'Success',
-        //     'Action' => 'Started'
-        // ));
-        // fastcgi_finish_request();
+        echo json_encode(array(
+            'Status' => 'Success',
+            'Action' => 'Started'
+        ));
+        fastcgi_finish_request();
 
         // Set Progress
         $Progress = 0;
@@ -66,7 +66,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         // Do Chart, Spreadsheet & Image Stuff Here ....
         // Top threat feeds
         $Progress = writeProgress($UUID,$Progress,"Building Threat Feeds");
-        //$TopThreatFeeds = QueryCubeJS('{"measures":["PortunusAggSecurity.feednameCount"],"dimensions":["PortunusAggSecurity.feed_name"],"timeDimensions":[{"dimension":"PortunusAggSecurity.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggSecurity.type","operator":"equals","values":["2"]},{"member":"PortunusAggSecurity.severity","operator":"equals","values":["High"]}],"limit":"10","ungrouped":false}');
         //print_r($CubeJSResults);
         $TopThreatFeeds = $CubeJSResults['TopThreatFeeds']['Body'];
         if (isset($TopThreatFeeds->result->data)) {
@@ -84,7 +83,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Top detected properties
         $Progress = writeProgress($UUID,$Progress,"Building Threat Properties");
-        //$TopDetectedProperties = QueryCubeJS('{"measures":["PortunusDnsLogs.tpropertyCount"],"dimensions":["PortunusDnsLogs.tproperty"],"timeDimensions":[{"dimension":"PortunusDnsLogs.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusDnsLogs.type","operator":"equals","values":["2"]},{"member":"PortunusDnsLogs.feed_name","operator":"notEquals","values":["Public_DOH","public-doh","Public_DOH_IP","public-doh-ip"]},{"member":"PortunusDnsLogs.severity","operator":"notEquals","values":["Low","Info"]}],"limit":"10","ungrouped":false}');
         $TopDetectedProperties = $CubeJSResults['TopDetectedProperties']['Body'];
         if (isset($TopDetectedProperties->result->data)) {
             $TopDetectedPropertiesSS = IOFactory::load($FilesDir.'/reports/report-'.$UUID.'/ppt/embeddings/Microsoft_Excel_Worksheet1.xlsx');
@@ -101,7 +99,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Content filtration
         $Progress = writeProgress($UUID,$Progress,"Building Content Filters");
-        //$ContentFiltration = QueryCubeJS('{"measures":["PortunusAggWebcontent.categoryCount"],"dimensions":["PortunusAggWebcontent.category"],"timeDimensions":[{"dimension":"PortunusAggWebcontent.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[],"limit":"10","ungrouped":false}');
         $ContentFiltration = $CubeJSResults['ContentFiltration']['Body'];
         if (isset($ContentFiltration->result->data)) {
             $ContentFiltrationSS = IOFactory::load($FilesDir.'/reports/report-'.$UUID.'/ppt/embeddings/Microsoft_Excel_Worksheet2.xlsx');
@@ -118,7 +115,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Insight Distribution by Threat Type - Sheet 3
         $Progress = writeProgress($UUID,$Progress,"Building SOC Insight Threat Types");
-        //$InsightDistribution = QueryCubeJS('{"measures":["InsightsAggregated.count"],"dimensions":["InsightsAggregated.threatType"],"filters":[{"member":"InsightsAggregated.insightStatus","operator":"equals","values":["Active"]}]}');
         $InsightDistribution = $CubeJSResults['InsightDistribution']['Body'];
         if (isset($InsightDistribution->result->data)) {
             $InsightDistributionSS = IOFactory::load($FilesDir.'/reports/report-'.$UUID.'/ppt/embeddings/Microsoft_Excel_Worksheet3.xlsx');
@@ -168,7 +164,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         // ** Reusable Metrics ** //
         // DNS Firewall Activity - Used on Slides 2, 5 & 6
         $Progress = writeProgress($UUID,$Progress,"Building DNS Firewall Event Criticality");
-        //$DNSFirewallActivity = QueryCubeJS('{"measures":["PortunusAggSecurity.severityCount"],"dimensions":["PortunusAggSecurity.severity"],"timeDimensions":[{"dimension":"PortunusAggSecurity.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggSecurity.type","operator":"equals","values":["2","3"]},{"member":"PortunusAggSecurity.severity","operator":"equals","values":["High","Medium","Low"]}],"limit":"3","ungrouped":false}');
         $DNSFirewallActivity = $CubeJSResults['DNSFirewallActivity']['Body'];
         if (isset($DNSFirewallActivity->result)) {
             $HighId = array_search('High', array_column($DNSFirewallActivity->result->data, 'PortunusAggSecurity.severity'));
@@ -195,7 +190,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Total DNS Activity - Used on Slides 6 & 9
         $Progress = writeProgress($UUID,$Progress,"Building DNS Activity");
-        //$DNSActivity = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["1"]}],"limit":"1","ungrouped":false}');
         $DNSActivity = $CubeJSResults['DNSActivity']['Body'];
         if (isset($DNSActivity->result->data[0])) {
             $DNSActivityCount = $DNSActivity->result->data[0]->{'PortunusAggInsight.requests'};
@@ -215,7 +209,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // SOC Insights - Used on Slides 15 & 28
         $Progress = writeProgress($UUID,$Progress,"Building SOC Insight Threat Criticality");
-        //$SOCInsights = QueryCubeJS('{"measures":["InsightsAggregated.count","InsightsAggregated.mostRecentAt","InsightsAggregated.startedAtMin"],"dimensions":["InsightsAggregated.priorityText"],"filters":[{"member":"InsightsAggregated.insightStatus","operator":"equals","values":["Active"]}],"timezone":"UTC"}');
         $SOCInsights = $CubeJSResults['SOCInsights']['Body'];
         if (isset($SOCInsights->result)) {
             $InfoInsightsId = array_search('INFO', array_column($SOCInsights->result->data, 'InsightsAggregated.priorityText'));
@@ -235,7 +228,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Security Activity
         $Progress = writeProgress($UUID,$Progress,"Building Security Activity");
-        //$SecurityEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"contains","values":["2","3"]}],"limit":"1","ungrouped":false}');
         $SecurityEvents = $CubeJSResults['SecurityEvents']['Body'];
         if (isset($SecurityEvents->result->data[0])) {
             $SecurityEventsCount = $SecurityEvents->result->data[0]->{'PortunusAggInsight.requests'};
@@ -245,7 +237,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Data Exfiltration Events
         $Progress = writeProgress($UUID,$Progress,"Building Data Exfiltration Events");
-        //$DataExfilEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["4"]},{"member":"PortunusAggInsight.tclass","operator":"equals","values":["TI-DNST"]}],"ungrouped":false}');
         $DataExfilEvents = $CubeJSResults['DataExfilEvents']['Body'];
         if (isset($DataExfilEvents->result->data[0])) {
             $DataExfilEventsCount = $DataExfilEvents->result->data[0]->{'PortunusAggInsight.requests'};
@@ -255,7 +246,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Zero Day DNS Events
         $Progress = writeProgress($UUID,$Progress,"Building Zero Day DNS Events");
-        //$ZeroDayDNSEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["2","3"]},{"member":"PortunusAggInsight.tclass","operator":"equals","values":["Zero Day DNS"]}],"ungrouped":false}');
         $ZeroDayDNSEvents = $CubeJSResults['ZeroDayDNSEvents']['Body'];
         if (isset($ZeroDayDNSEvents->result->data[0])) {
             $ZeroDayDNSEventsCount = $ZeroDayDNSEvents->result->data[0]->{'PortunusAggInsight.requests'};
@@ -265,7 +255,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Suspicious Domains
         $Progress = writeProgress($UUID,$Progress,"Building Suspicious Domain Events");
-        //$SuspiciousEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["2"]},{"member":"PortunusAggInsight.tclass","operator":"equals","values":["Suspicious"]}],"ungrouped":false}');
         $SuspiciousEvents = $CubeJSResults['SuspiciousEvents']['Body'];
         if (isset($SuspiciousEvents->result->data[0])) {
             $SuspiciousEventsCount = $SuspiciousEvents->result->data[0]->{'PortunusAggInsight.requests'};
@@ -275,7 +264,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // High Risk Websites
         $Progress = writeProgress($UUID,$Progress,"Building High Risk Website Events");
-        //$HighRiskWebsites = QueryCubeJS('{"timeDimensions":[{"dimension":"PortunusAggWebContentDiscovery.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"measures":["PortunusAggWebContentDiscovery.count","PortunusAggWebContentDiscovery.deviceCount"],"dimensions":["PortunusAggWebContentDiscovery.domain_category"],"order":{"PortunusAggWebContentDiscovery.count":"desc"},"filters":[{"member":"PortunusAggWebContentDiscovery.domain_category","operator":"equals","values":["Risky Activity","Suspicious and Malicious Software","Uncategorized","Adult","Abortion","Abortion Pro Choice","Abortion Pro Life","Child Inappropriate","Gambling","Gay","Lingerie","Nudity","Pornography","Profanity","R-Rated","Sex & Erotic","Sex Education","Tobacco","Anonymizer","Criminal Skills","Self Harm","Criminal Activities - Other","Illegal Drugs","Marijuana","Child Abuse Images","Hacking","Hate Speech","Piracy & Copyright Theft","Torrent Repository","Terrorism","Peer-to-Peer","Violence","Weapons","School Cheating","Ad Fraud","Botnet","Command and Control Centers","Compromised & Links To Malware","Malware Call-Home","Malware Distribution Point","Phishing/Fraud","Spam URLs","Spyware & Questionable Software","Cryptocurrency Mining","Sexuality","Parked & For Sale Domains"]}]}');
         $HighRiskWebsites = $CubeJSResults['HighRiskWebsites']['Body'];
         if (isset($HighRiskWebsites->result->data)) {
             $HighRiskWebsiteCount = array_sum(array_column($HighRiskWebsites->result->data, 'PortunusAggWebContentDiscovery.count'));
@@ -287,7 +275,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // DNS over HTTPS
         $Progress = writeProgress($UUID,$Progress,"Building DoH Events");
-        //$DOHEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["2"]},{"member":"PortunusAggInsight.tproperty","operator":"equals","values":["DoHService"]}],"ungrouped":false}');
         $DOHEvents = $CubeJSResults['DOHEvents']['Body'];
         if (isset($DOHEvents->result->data[0])) {
             $DOHEventsCount = $DOHEvents->result->data[0]->{'PortunusAggInsight.requests'};
@@ -297,7 +284,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Newly Observed Domains
         $Progress = writeProgress($UUID,$Progress,"Building Newly Observed Domain Events");
-        //$NODEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["2"]},{"member":"PortunusAggInsight.tproperty","operator":"equals","values":["NewlyObservedDomains"]}],"ungrouped":false}');
         $NODEvents = $CubeJSResults['NODEvents']['Body'];
         if (isset($NODEvents->result->data[0])) {
             $NODEventsCount = $NODEvents->result->data[0]->{'PortunusAggInsight.requests'};
@@ -307,7 +293,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Domain Generation Algorithms
         $Progress = writeProgress($UUID,$Progress,"Building DGA Events");
-        //$DGAEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"or":[{"member":"PortunusAggInsight.tproperty","operator":"equals","values":["suspicious_rdga","DGA"]},{"member":"PortunusAggInsight.tclass","operator":"equals","values":["DGA","MalwareC2DGA"]}]},{"member":"PortunusAggInsight.type","operator":"equals","values":["2","3"]}],"ungrouped":false}');
         $DGAEvents = $CubeJSResults['DGAEvents']['Body'];
         if (isset($DGAEvents->result->data[0])) {
             $DGAEventsCount = $DGAEvents->result->data[0]->{'PortunusAggInsight.requests'};
@@ -317,7 +302,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Unique Applications
         $Progress = writeProgress($UUID,$Progress,"Building list of Unique Applications");
-        //$UniqueApplications = QueryCubeJS('{"measures":["PortunusAggAppDiscovery.requests"],"dimensions":["PortunusAggAppDiscovery.app_name","PortunusAggAppDiscovery.app_approval"],"timeDimensions":[{"dimension":"PortunusAggAppDiscovery.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggAppDiscovery.app_name","operator":"set"},{"member":"PortunusAggAppDiscovery.app_name","operator":"notEquals","values":[""]}],"order":{}}');
         $UniqueApplications = $CubeJSResults['UniqueApplications']['Body'];
         if (isset($UniqueApplications->result->data)) {
             $UniqueApplicationsCount = count($UniqueApplications->result->data);
@@ -345,7 +329,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Threat Activity
         $Progress = writeProgress($UUID,$Progress,"Building Threat Activity");
-        //$ThreatActivityEvents = QueryCubeJS('{"measures":["PortunusAggInsight.threatCount"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["2"]},{"member":"PortunusAggInsight.severity","operator":"equals","values":["High","Medium","Low"]},{"member":"PortunusAggInsight.threat_indicator","operator":"notEquals","values":[""]}],"limit":"1","ungrouped":false}');
         $ThreatActivityEvents = $CubeJSResults['ThreatActivityEvents']['Body'];
         if (isset($ThreatActivityEvents->result->data[0])) {
             $ThreatActivityEventsCount = $ThreatActivityEvents->result->data[0]->{'PortunusAggInsight.threatCount'};
@@ -355,7 +338,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // DNS Firewall
         $Progress = writeProgress($UUID,$Progress,"Building DNS Firewall Activity");
-        //$DNSFirewallEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"and":[{"member":"PortunusAggInsight.type","operator":"equals","values":["2"]},{"or":[{"member":"PortunusAggInsight.severity","operator":"equals","values":["High","Medium","Low"]},{"and":[{"member":"PortunusAggInsight.severity","operator":"equals","values":["Info"]},{"member":"PortunusAggInsight.policy_action","operator":"equals","values":["Block","Log"]}]}]},{"member":"PortunusAggInsight.confidence","operator":"equals","values":["High","Medium","Low"]}]}],"limit":"1","ungrouped":false}');
         $DNSFirewallEvents = $CubeJSResults['DNSFirewallEvents']['Body'];
         if (isset($DNSFirewallEvents->result->data[0])) {
             $DNSFirewallEventsCount = $DNSFirewallEvents->result->data[0]->{'PortunusAggInsight.requests'};
@@ -365,7 +347,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Web Content
         $Progress = writeProgress($UUID,$Progress,"Building Web Content Events");
-        //$WebContentEvents = QueryCubeJS('{"measures":["PortunusAggWebcontent.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggWebcontent.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggWebcontent.type","operator":"equals","values":["3"]},{"member":"PortunusAggWebcontent.category","operator":"notEquals","values":[null]}],"limit":"1","ungrouped":false}');
         $WebContentEvents = $CubeJSResults['WebContentEvents']['Body'];
         if (isset($WebContentEvents->result->data[0])) {
             $WebContentEventsCount = $WebContentEvents->result->data[0]->{'PortunusAggWebcontent.requests'};
@@ -375,7 +356,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Device Count
         $Progress = writeProgress($UUID,$Progress,"Building Device Count");
-        //$Devices = QueryCubeJS('{"measures":["PortunusAggInsight.deviceCount"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"contains","values":["2","3"]},{"member":"PortunusAggInsight.severity","operator":"contains","values":["High","Medium","Low"]}],"limit":"1","ungrouped":false}');
         $Devices = $CubeJSResults['Devices']['Body'];
         if (isset($Devices->result->data[0])) {
             $DeviceCount = $Devices->result->data[0]->{'PortunusAggInsight.deviceCount'};
@@ -385,7 +365,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // User Count
         $Progress = writeProgress($UUID,$Progress,"Building User Count");
-        //$Users = QueryCubeJS('{"measures":["PortunusAggInsight.userCount"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"contains","values":["2","3"]}],"limit":"1","ungrouped":false}');
         $Users = $CubeJSResults['Users']['Body'];
         if (isset($Users->result->data[0])) {
             $UserCount = $Users->result->data[0]->{'PortunusAggInsight.userCount'};
@@ -395,7 +374,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Threat Insight Count
         $Progress = writeProgress($UUID,$Progress,"Building Threat Insight Count");
-        //$ThreatInsight = QueryCubeJS('{"measures":[],"dimensions":["PortunusDnsLogs.tproperty"],"timeDimensions":[{"dimension":"PortunusDnsLogs.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusDnsLogs.type","operator":"equals","values":["4"]}],"limit":"10000","ungrouped":false}'); // Threat Insight
         $ThreatInsight = $CubeJSResults['ThreatInsight']['Body'];
         if (isset($ThreatInsight->result->data)) {
             $ThreatInsightCount = count($ThreatInsight->result->data);
@@ -405,7 +383,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Threat View Count
         $Progress = writeProgress($UUID,$Progress,"Building Threat View Count");
-        //$ThreatView = QueryCubeJS('{"measures":["PortunusAggInsight.tpropertyCount"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["2"]}],"limit":"1","ungrouped":false}');
         $ThreatView = $CubeJSResults['ThreatView']['Body'];
         if (isset($ThreatView->result->data[0])) {
             $ThreatViewCount = $ThreatView->result->data[0]->{'PortunusAggInsight.tpropertyCount'};
@@ -415,7 +392,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Source Count
         $Progress = writeProgress($UUID,$Progress,"Building Sources Count");
-        //$Sources = QueryCubeJS('{"measures":["PortunusAggSecurity.networkCount"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggSecurity.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggSecurity.type","operator":"contains","values":["2","3"]}],"limit":"1","ungrouped":false}');
         $Sources = $CubeJSResults['Sources']['Body'];
         if (isset($Sources->result->data[0])) {
             $SourcesCount = $Sources->result->data[0]->{'PortunusAggSecurity.networkCount'};
@@ -424,7 +400,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         }
 
         // ** ** //
-
 
         //
         // Do Threat Actor Stuff Here ....
@@ -440,7 +415,7 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
 
         // Open PPTX Presentation _rels XML
         $xml_rels = new DOMDocument('1.0', 'utf-8');
-        $xml_rels->formatOutput = true; 
+        $xml_rels->formatOutput = true;
         $xml_rels->preserveWhiteSpace = false;
         $xml_rels->load($FilesDir.'/reports/report-'.$UUID.'/ppt/_rels/presentation.xml.rels');
         $xml_rels_f = $xml_rels->createDocumentFragment();
@@ -505,7 +480,7 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
             // Update Slide XML with changes
             file_put_contents($FilesDir.'/reports/report-'.$UUID.'/ppt/slides/slide'.$SlideNumber.'.xml', $TASFile);
             $xml_tas = new DOMDocument('1.0', 'utf-8');
-            $xml_tas->formatOutput = true; 
+            $xml_tas->formatOutput = true;
             $xml_tas->preserveWhiteSpace = false;
             $xml_tas->load($FilesDir.'/reports/report-'.$UUID.'/ppt/slides/_rels/slide'.$SlideNumber.'.xml.rels');
             foreach ($xml_tas->getElementsByTagName('Relationship') as $element) {
@@ -561,15 +536,15 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         $xml_rels->getElementsByTagName('Relationships')->item(0)->appendChild($xml_rels_f);
         // Append new slides to specific position
         $xml_pres->getElementsByTagName('sldId')->item($ThreatActorSlidePosition)->after($xml_pres_f);
-        
+
         // Save Core XML Files
         $xml_rels->save($FilesDir.'/reports/report-'.$UUID.'/ppt/_rels/presentation.xml.rels');
         $xml_pres->save($FilesDir.'/reports/report-'.$UUID.'/ppt/presentation.xml');
-        
+
         //
         // End of Threat Actors
         //
-       
+
 
         // Rebuild Powerpoint Template Zip
         $Progress = writeProgress($UUID,$Progress,"Stitching Powerpoint Template");
@@ -788,10 +763,17 @@ function getProgress($id,$Total) {
     if (file_exists($ProgressFile)) {
         $myfile = fopen($ProgressFile, "r") or die("0");
         $Current = json_decode(fread($myfile,filesize($ProgressFile)));
-        return array(
-            'Progress' => (100 / $Total) * $Current->Count,
-            'Action' => $Current->Action.'..'
-        );
+        if (isset($Current) && isset($Current->Count) && isset($Current->Action)) {
+            return array(
+                'Progress' => (100 / $Total) * $Current->Count,
+                'Action' => $Current->Action.'..'
+            );
+        } else {
+            return array(
+                'Progress' => 0,
+                'Action' => 'Checking..'
+            );
+        }
     } else {
         return array(
             'Progress' => 0,