+++ inc/inc.php
@@ -1,6 +1,7 @@
 <?php
 // Set Max Script Execution Time
 ini_set('max_execution_time', '300');
+ini_set('error_log',__DIR__.'/nginx/php.error.log');
 // Include Functions
 require_once(__DIR__.'/../vendor/autoload.php');
 require_once(__DIR__.'/functions/auth-functions.php');